extractor/extractor/utils.py
from difflib import SequenceMatcher
from google.cloud import storage

def download_blob(bucket_name, file_name):
    storage_client = storage.Client()
    bucket = storage_client.bucket(bucket_name)
    blob = bucket.blob(file_name)
    return blob.download_as_text()


