extractor/pyproject.toml
[project]
name = "extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=2.19.0",
    "google-cloud-videointelligence>=2.15.0",
    "numpy>=2.2.1",
    "pydantic>=2.10.4",
    "themoviedb[async]>=1.0.2",
    "yt-dlp>=2024.12.23",
]

[project.scripts]
hello = "extractor:main"

[dependency-groups]
dev = [
    "flask>=3.1.0",
    "functions-framework>=3.8.2",
    "ipykernel>=6.29.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.google-cloud-function]
path = "src/extractor/main.py"
dest = "main.py"
